requirement.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
fastapi==0.115.7
fonttools==4.55.5
geojson==3.2.0
h11==0.14.0
idna==3.10
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
networkx==3.4.2
numpy==2.2.2
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
scipy==1.15.1
shapely==2.0.6
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
